Ongoing refactoring.
