+++ app/src/main/java/com/bogdan/codeforceswatcher/feature/contests/redux/request/ContestResponse.kt
@@ -1,4 +1,4 @@
-package com.bogdan.codeforceswatcher.feature.contests
+package com.bogdan.codeforceswatcher.feature.contests.redux.request
 
 import com.bogdan.codeforceswatcher.model.Contest
