+++ app/src/main/java/com/bogdan/codeforceswatcher/room/Database.kt
@@ -3,7 +3,7 @@ package com.bogdan.codeforceswatcher.room
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.bogdan.codeforceswatcher.feature.contests.ContestDao
+import com.bogdan.codeforceswatcher.feature.contests.room.ContestDao
 import com.bogdan.codeforceswatcher.model.Contest
 import com.bogdan.codeforceswatcher.model.User
